@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe MuseumsController do
 
@@ -12,8 +12,8 @@ describe MuseumsController do
       csv_to_import = fixture_file_upload(Rails.root.join('spec', 'fixtures', 'museums.csv'), "text/csv")
       post :import, file: csv_to_import
 
-      flash[:notice].should eq("Museums imported successfully.")
-      Museum.count.should == 5
+      expect(flash[:notice]).to eq("Museums imported successfully.")
+      expect(Museum.count).to eq(5)
     end
   end
 
@@ -33,8 +33,8 @@ describe MuseumsController do
 
     it "should create new museum" do
       get :new
-      response.should be_ok
-      assigns[:museum].should be_present
+      expect(response).to be_ok
+      expect(assigns[:museum]).to be_present
     end
   end
 
@@ -46,8 +46,8 @@ describe MuseumsController do
     end
 
     it "should redirect to show page for museum with create success message" do
-      subject.should redirect_to assigns(:museum)
-      flash[:notice].should eq("Museum was successfully created.")
+      expect(subject).to redirect_to(assigns(:museum))
+      expect(flash[:notice]).to eq("Museum was successfully created.")
     end
   end
 
@@ -55,7 +55,7 @@ describe MuseumsController do
     it "should return success" do
       museum = FactoryGirl.create(:museum)
       get :show, id: museum.id
-      response.should be_ok
+      expect(response).to be_ok
     end
   end
 
@@ -63,7 +63,7 @@ describe MuseumsController do
     it "should return success" do
       museum = FactoryGirl.create(:museum)
       get :edit, id: museum.id
-      response.should be_ok
+      expect(response).to be_ok
     end
   end
 
@@ -72,10 +72,9 @@ describe MuseumsController do
       museum = FactoryGirl.create(:museum)
       patch :update, id: museum.id, museum: FactoryGirl.attributes_for(:museum, name:"A New Museum Name")
       museum.reload
-      museum.name.should eq("A New Museum Name")
-      museum.slug.should eq("a-new-museum-name")
-      response.should redirect_to assigns(:museum)
-
+      expect(museum.name).to eq("A New Museum Name")
+      expect(museum.slug).to eq("a-new-museum-name")
+      expect(response).to redirect_to(assigns(:museum))
     end
   end
 
