@@ -1,13 +1,13 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe ApplicationHelper do
 	describe "boolean_to_human" do
 		it "should return 'Yes' when passed true" do
-			boolean_to_human(true).should eq("Yes")
+      expect(boolean_to_human(true)).to eq("Yes")
 		end
 		
 		it "should return 'No' when passed false" do
-			boolean_to_human(false).should eq("No")
+      expect(boolean_to_human(false)).to eq("No")
 		end
 	end
 end
