@@ -1,13 +1,6 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe FamilyCard do
-  it { should validate_presence_of(:pass_id) }
-  it { should validate_presence_of(:first_name) }
-  it { should validate_presence_of(:last_name) }
-  it { should validate_presence_of(:organization_name) }
-
-  it { should validate_uniqueness_of(:pass_id) }
-
   describe ".import" do
     before :each do
       @file = Tempfile.new('users.csv')
@@ -25,13 +18,13 @@ describe FamilyCard do
 
     it 'should create new users when a valid CSV is imported' do
       FamilyCard.import(@file)
-      FamilyCard.count.should eq 10
+      expect(FamilyCard.count).to eq(10)
     end
 
     it 'should associate users and CSVs correctly' do
       FamilyCard.import(@file)
       family_card = FamilyCard.find(10001)
-      family_card.user.email.should eq "temp1@gmail.com"
+      expect(family_card.user.email).to eq("temp1@gmail.com")
     end
 
     it 'should not allow duplicates' do
@@ -43,7 +36,7 @@ describe FamilyCard do
       FamilyCard.import(@file)
 
       errors = FamilyCard.import(same_users)
-      errors.count.should eq 1
+      expect(errors.count).to eq(1)
       
       same_users.unlink
     end
